Restructure on disk
As we add support for the module unification layout the project will
likely end up with two more resolvers:

* addons/resolvers/module-unification
* addons/resolvers/migration

Pave the path for that. Additionally, drop the `create` polyfill
assertion as this library only supports IE9+ where Object.create is
present.
