+++ addon/utils/make-dictionary.js
@@ -1,7 +1,5 @@
-import create from './create';
-
 export default function makeDictionary() {
-  let cache = create(null);
+  let cache = Object.create(null);
   cache['_dict'] = null;
   delete cache['_dict'];
   return cache;
